+++ R/nepal_sees_jags_output.R
@@ -9,11 +9,12 @@
 #' paratyphoid). Keeping only IDs `"newperson"`, `"sees_npl_1"`, `"sees_npl_2"`.
 #'
 #' @format An S3 object of class `sr_model`: A [tibble::tbl_df] that contains 
-#' the posterior predictive distribution of the person-specific parameters for 
-#' a "new person" with no observed data (`Subject = "newperson"`) and posterior
-#' distributions of the person-specific parameters for two arbitrarily-chosen
-#' subjects (`"sees_npl_1"` and `"sees_npl_2"`).
-#' Contains 40,000 `rows`, 7 `columns`, and model `attributes`.
+#'   the
+#'   posterior predictive distribution of the person-specific parameters for a
+#'   "new person" with no observed data (`Subject = "newperson"`) and posterior
+#'   distributions of the person-specific parameters for two arbitrarily-chosen
+#'   subjects (`"sees_npl_1"` and `"sees_npl_2"`).
+#'   Contains 40,000 `rows`, 7 `columns`, and model `attributes`.
 #' \describe{
 #'  \item{Iteration}{Number of sampling iterations: 500 iterations}
 #'  \item{Chain}{Number of MCMC chains run: 2 chains run}