+++ NEWS.md
@@ -47,6 +47,7 @@
 
 ## New features
 
+* Including optional population parameters as attributes in run_mod output. (#141)
 * Made "newperson" optional in `prep_data()` (#73)
 * Including fitted and residual values as data frame in run_mod output. (#101)
 * Added  `plot_predicted_curve()` with support for faceting by multiple IDs (#68)